Add 'version' command.
It was frustrating to see the gvm version every time I ran a command, so I
have removed the relevant echo and moved it into a separate 'version' command.

I also decided to change the logic for determining whether the given command
is valid or not. The original check was already unwieldy and getting more so,
so the script now checks the given command against all the functions starting
with 'gvm-<cmd>'. If the function exists, so does the command.