Fix cucumber tests for per-shell versions feature.
Since gvm is now a function that is sourced into the current shell, the old
Cucumber tests no longer work as they try to execute 'gvm' as a process. This
fix instead launches an external bash process for each scenario and pipes the
required commands to it.

A lot of the magic happens in the new BashEnv class which handles the piping
of commands and also managing the output correctly.

This is further work on issue #52.