Fix problem switching from local version to standard.
In isolated mode, attempting to switch from a locally installed version to a
standard one would fail if the label for the local version did not start with
a number, e.g. 'dev'. One of the regular expressions in gvm-use.sh was still
assuming that versions always began with a number, which of course is no longer
the case.

I fixed the regular expression and added a corresponding Cucumber scenario.