Remove hard-coded references to candidates
It seems wrong to hard-code the candidate names when they are
available from the server.  It also makes it hard to add new
candidates.  The scripts pretty much work with a really small
change to gvm-use.sh, but the tests also rely on the PATH being
set, so there is a change to the tests as well to make them pass.