Add some hooks, searching in GVM_DIR/hooks
E.g. for grails 2.1.0 the system looks for a
post-install.sh in

hooks/grails/2.1.0
hooks/grails/2.1
hooks/grails/2
hooks/grails/default
hooks

It executes the first one it finds and then
exits the loop.

N.B. this patch does not install any hooks.  We
need to decide how that could be done in a data-driven
way.  Registering a hooks.zip per candidate
in Mongo might be one option, and unzipping it
during an installation.