Dynamically reinitialize sdkman
Check whether a candidate is on the path each time sdkman_init.sh is
sourced. If the candidate is not on the path prefix it, otherwise do
nothing.

This should resolve #251 and resolve #299

Since the path is dynamically checked now the SDKMAN_INIT flag and test
are no longer necessary