If `less` is not present, just output directly.
Sometimes `less` isn't available—for example, in the `java:8` Docker
image. If it's not there and there's no `$PAGER` variable is set, it's
best to just output directly.

Unfortunately, I couldn't figure out a way to test this without renaming
`less` to something else and then renaming it back afterwards, which
seemed… dangerous, so I've left it untested. Ideas are definitely
welcome.