Refine bash major version check
The previous bash version check, `bash --version | grep ... `, may not work when the output is not in English (for example, it didn't work when the output was in Italian). For the same reason, I've modified also the line 113, to output the first command output of `bash --version` without checking for the content.