Use '>|' to write var/ status files
Setting 'noclobber' in bash or zsh prevents writing to an existing
regular file when using > as redirecting operator. Setting 'noclobber' allows
users of interactive shells to prevent accidential overwriting of
existing files.

bash and zsh allow to use >| as redirection operator so that the
redirection is attempted.

Because the write to the status files is required and not by accident
the >| is used in those cases.