Use shell-agnostic code in …_secure_curl_download
curl_params in __sdkman_secure_curl_download is now an array,
so the function no longer requires bash- & zsh-specific code.

The order of the arguments to curl has not been preserved by this commit
in order to make the git diff more understandable. The next commit
rearranges the if conditionals to preserve the original argument order.

Also fixes argument concatenation issue in bash.