Add GitHub issue templates for bugs, features, support and PRs
Issues are either a bug report, feature request or support request. Labels
are automaticaly added to the issues respectively: bug, enhancement, support.

The templates suggest using the Slack before opening an issue, but no longer
say it is a hard requirement. This might give a more inclusive impression to
people reporting who are not Slack users.